Further work: start wiring up servant.
The DB transaction is still very confusing to me. How does
interpretation work with transaction support? How do we run
transactions across the connections from the pool (but one transaction
needs to run on a single connection?).

Is it just better to use a simpler and less confusing approach at this time?